# src/app/cart/dto/cart-product.dto.ts
export class CartProductDto {
    public cartId: number;
    public attribute: string;
    public quantity: number;
    public name: string;
    public image: string;
    public subtotal: number;
    public price: number;
    public productId: number;
}