# swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Backend Challenge",
        "version": "1.0.0",
        "title": "Turing/E-commerce",
        "contact": {
            "email": "codeninja0x01@gmail.com"
        }
    },
    "basePath": "/api",
...
    "schemes": [
        "http"
    ],
    "securityDefinitions": {
        "bearer": {
            "type": "apiKey",
            "name": "USER-KEY",
            "in": "header"
        }
    },
    "host": "localhost:3000",
    "paths": {
...
        "/attributes/{id}": {
            "get": {
                "summary": "Get Attribute.",
                "description": "Return a single attribute.",
                "parameters": [
                    {
                        "type": "number",
                        "name": "id",
                        "required": true,
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successfully operation.",
                        "schema": {
                            "$ref": "#/definitions/Attribute"
                        }
                    },
                    "422": {
                        "description": "Invalid user input."
                    }
                },
                "tags": [
                    "attribute"
                ],
                "produces": [
                    "application/json"
                ],
                "consumes": [
                    "application/json"
                ]
            }
        },
        "/attributes/value/{id}": {
            "get": {
                "summary": "Get Attribute Value.",
                "description": "Return a list of attributes value.",
                "parameters": [
                    {
                        "type": "number",
                        "name": "id",
                        "required": true,
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successfully operation.",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/AttributeValue"
                            }
                        }
                    },
                    "422": {
                        "description": "Invalid user input."
                    }
                },
                "tags": [
                    "attribute"
                ],
                "produces": [
                    "application/json"
                ],
                "consumes": [
                    "application/json"
                ]
            }
        },
        "/attributes/product/{productId}": {
            "get": {
                "summary": "Get Attribute product.",
                "description": "Return a list of attribute product.",
                "parameters": [
                    {
                        "type": "number",
                        "name": "productId",
                        "required": true,
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successfully operation.",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/AttributeProduct"
                            }
                        }
                    },
                    "422": {
                        "description": "Invalid user input."
                    }
                },
                "tags": [
                    "attribute"
                ],
                "produces": [
                    "application/json"
                ],
                "consumes": [
                    "application/json"
                ]
            }
        },
        "/carts": {
            "post": {
                "summary": "Add cart.",
                "description": "Create cart and return cart product.",
                "parameters": [
                    {
                        "name": "attribute",
                        "required": true,
                        "in": "formData",
                        "type": "string"
                    },
                    {
                        "name": "cartKey",
                        "required": true,
                        "in": "formData",
                        "type": "string"
                    },
                    {
                        "name": "productId",
                        "required": true,
                        "in": "formData",
                        "type": "number"
                    },
                    {
                        "name": "quantity",
                        "required": true,
                        "in": "formData",
                        "type": "number"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "successfully operation.",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/CartProduct"
                            }
                        }
                    },
                    "400": {
                        "description": "Input data validation failed."
                    }
                },
                "tags": [
                    "cart"
                ],
                "produces": [
                    "application/json"
                ],
                "consumes": [
                    "application/x-www-form-urlencoded"
                ]
            }
        },
        "/carts/{id}": {
            "post": {
                "summary": "Update cart.",
                "description": "Update cart quantity and return cart product.",
                "parameters": [
                    {
                        "type": "number",
                        "name": "id",
                        "required": true,
                        "in": "path"
                    },
                    {
                        "name": "quantity",
                        "required": true,
                        "in": "formData",
                        "type": "number"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "successfully operation.",
                        "schema": {
                            "$ref": "#/definitions/CartProduct"
                        }
                    },
                    "400": {
                        "description": "Input data validation failed."
                    }
                },
                "tags": [
                    "cart"
                ],
                "produces": [
                    "application/json"
                ],
                "consumes": [
                    "application/x-www-form-urlencoded"
                ]
            }
        },
...
        "/carts/{cartKey}": {
            "get": {
                "summary": "Get cart product.",
                "description": "Return cart product.",
                "parameters": [
                    {
                        "type": "string",
                        "name": "cartKey",
                        "required": true,
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successfully operation.",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/CartProduct"
                            }
                        }
                    },
                    "400": {
                        "description": "Input data validation failed."
                    }
                },
                "tags": [
                    "cart"
                ],
                "produces": [
                    "application/json"
                ],
                "consumes": [
                    "application/json"
                ]
            }
        },
        "/carts/{id}/buynow": {
            "get": {
                "summary": "Buy product now.",
                "description": "Move product to cart.",
                "parameters": [
                    {
                        "type": "number",
                        "name": "id",
                        "required": true,
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successfully operation."
                    },
                    "400": {
                        "description": "Input data validation failed."
                    }
                },
                "tags": [
                    "cart"
                ],
                "produces": [
                    "application/json"
                ],
                "consumes": [
                    "application/json"
                ]
            }
        },
        "/carts/{id}/buylater": {
            "get": {
                "summary": "Buy product later.",
                "description": "Save product for later.",
                "parameters": [
                    {
                        "type": "number",
                        "name": "id",
                        "required": true,
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successfully operation."
                    },
                    "400": {
                        "description": "Input data validation failed."
                    }
                },
                "tags": [
                    "cart"
                ],
                "produces": [
                    "application/json"
                ],
                "consumes": [
                    "application/json"
                ]
            }
        },
        "/carts/{cartKey}/saved": {
            "get": {
                "summary": "Get saved product.",
                "description": "Return saved product.",
                "parameters": [
                    {
                        "type": "string",
                        "name": "cartKey",
                        "required": true,
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successfully operation.",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/CartProduct"
                            }
                        }
                    },
                    "400": {
                        "description": "Input data validation failed."
                    }
                },
                "tags": [
                    "cart"
                ],
                "produces": [
                    "application/json"
                ],
                "consumes": [
                    "application/json"
                ]
            }
        },
        "/carts/{id}/total": {
            "get": {
                "summary": "Get total cart price.",
                "description": "Return total cart product price.",
                "parameters": [
                    {
                        "type": "number",
                        "name": "id",
                        "required": true,
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successfully operation."
                    },
                    "400": {
                        "description": "Input data validation failed."
                    }
                },
                "tags": [
                    "cart"
                ],
                "produces": [
                    "application/json"
                ],
                "consumes": [
                    "application/json"
                ]
            }
        },
        "/carts/empty/{cartKey}": {
            "delete": {
                "summary": "Remove cart.",
                "description": "Delete cart.",
                "parameters": [
                    {
                        "type": "string",
                        "name": "cartKey",
                        "required": true,
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successfully operation."
                    },
                    "400": {
                        "description": "Input data validation failed."
                    }
                },
                "tags": [
                    "cart"
                ],
                "produces": [
                    "application/json"
                ],
                "consumes": [
                    "application/json"
                ]
            }
        },
        "/carts/product/{id}": {
            "delete": {
                "summary": "Remove product.",
                "description": "Delete product from cart.",
                "parameters": [
                    {
                        "type": "number",
                        "name": "id",
                        "required": true,
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successfully operation."
                    },
                    "400": {
                        "description": "Input data validation failed."
                    }
                },
                "tags": [
                    "cart"
                ],
                "produces": [
                    "application/json"
                ],
                "consumes": [
                    "application/json"
                ]
            }
        },
...
    },
...
    "definitions": {
...
        "Attribute": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "number"
                },
                "name": {
                    "type": "string"
                }
            },
            "required": [
                "id",
                "name"
            ]
        },
        "AttributeValue": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "number"
                },
                "name": {
                    "type": "string"
                }
            },
            "required": [
                "id",
                "name"
            ]
        },
        "AttributeProduct": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                }
            },
            "required": [
                "id",
                "name",
                "value"
            ]
        },
        "CartProduct": {
            "type": "object",
            "properties": {
                "cartId": {
                    "type": "number"
                },
                "attribute": {
                    "type": "string"
                },
                "quantity": {
                    "type": "number"
                },
                "name": {
                    "type": "string"
                },
                "image": {
                    "type": "string"
                },
                "subtotal": {
                    "type": "number"
                },
                "price": {
                    "type": "number"
                },
                "productId": {
                    "type": "number"
                }
            },
            "required": [
                "cartId",
                "attribute",
                "quantity",
                "name",
                "image",
                "subtotal",
                "price",
                "productId"
            ]
        },
...
    }
}